Add logger
